rename: change top component to header component
