+++ src/components/common/index.ts
@@ -1,6 +1,6 @@
 export * from "./icon"
 export * from "./text"
 export * from "./tooltip"
-export * from "./top/top"
-export * from "./top/top-back"
+export * from "./header/header"
+export * from "./header/header-back"
